refactor: eliminate sqlite.py re-export shim, relocate functions to proper modules
- Delete polylogue/storage/backends/sqlite.py (backward-compat re-export hub)
- Move DatabaseError → errors.py (joins error hierarchy)
- Move _parse_json, _row_get, _row_to_* → store.py (near record types)
- Move _iso_to_epoch, _build_conversation_filters → connection.py (query layer)
- Redirect ~100 import sites across production and test code
- Update mock.patch() strings and importlib.reload() targets
- Rename async_build() → build() in ConversationBuilder (no sync counterpart)
- Remove stale "backward compatibility" comments from config.py, connection.py, event_bus.py
- Fix importlib.reload ordering in test_store_ops.py (clear cache before reload)

46 files changed, 4213 tests pass.
