+++ tests/unit/cli/test_check.py
@@ -9,7 +9,7 @@ from click.testing import CliRunner
 
 from polylogue.cli import cli
 from polylogue.health import HealthCheck, HealthReport, VerifyStatus
-from polylogue.storage.backends.sqlite import open_connection
+from polylogue.storage.backends.connection import open_connection
 from tests.infra.helpers import DbFactory
