+++ tests/unit/core/test_filters_adv.py
@@ -19,7 +19,7 @@ import pytest
 from polylogue.lib.filters import ConversationFilter
 from polylogue.lib.models import ConversationSummary
 from polylogue.storage.backends.async_sqlite import SQLiteBackend
-from polylogue.storage.backends.sqlite import open_connection
+from polylogue.storage.backends.connection import open_connection
 from polylogue.storage.repository import ConversationRepository
 from polylogue.storage.index import rebuild_index
 from tests.infra.helpers import ConversationBuilder
