+++ tests/unit/pipeline/test_async_helpers.py
@@ -15,7 +15,7 @@ from uuid import uuid4
 import pytest
 
 from polylogue.storage.backends.async_sqlite import SQLiteBackend
-from polylogue.storage.backends.sqlite import open_connection
+from polylogue.storage.backends.connection import open_connection
 from polylogue.storage.store import ConversationRecord, MessageRecord
