+++ tests/unit/pipeline/test_prepare.py
@@ -6,7 +6,7 @@ from polylogue.export import export_jsonl
 from polylogue.paths import is_within_root
 from polylogue.rendering.renderers import HTMLRenderer
 from polylogue.sources import RecordBundle, save_bundle
-from polylogue.storage.backends.sqlite import open_connection
+from polylogue.storage.backends.connection import open_connection
 from tests.infra.helpers import make_attachment, make_conversation, make_message
